--- a/door/model/door.tres
+++ b/door/model/door.tres
@@ -7,6 +7,7 @@
 
 [resource]
 resource_name = "door"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 4 )
 metallic = 1.0
 metallic_texture = ExtResource( 1 )