--- a/door/model/door_unlockedemission.tres
+++ b/door/model/door_unlockedemission.tres
@@ -2,5 +2,6 @@
 
 [resource]
 resource_name = "door_unlockedemission"
+params_diffuse_mode = 1
 albedo_color = Color( 0.00685012, 0.64, 0, 1 )
 roughness = 0.0960784