--- a/enemies/red_robot/laser/impact_effect/EmissionEmbers.shader
+++ b/enemies/red_robot/laser/impact_effect/EmissionEmbers.shader
@@ -1,5 +1,5 @@
 shader_type spatial;
-render_mode blend_mix, depth_draw_opaque, cull_back, diffuse_burley, specular_schlick_ggx, cull_disabled;
+render_mode blend_mix, depth_draw_opaque, cull_back, diffuse_lambert, specular_disabled, cull_disabled;
 uniform vec4 albedo : hint_color;
 uniform sampler2D texture_albedo : hint_albedo;
 uniform float emission_energy;