--- a/enemies/red_robot/red_robot.tscn
+++ b/enemies/red_robot/red_robot.tscn
@@ -400,7 +400,7 @@ nodes/hit3/position = Vector2( 980, 240 )
 nodes/output/position = Vector2( 1180, 200 )
 nodes/state/node = SubResource( 51 )
 nodes/state/position = Vector2( 60, 100 )
-node_connections = [ "state", 0, "Animation", "state", 1, "Animation 3", "state", 2, "Animation 2", "state", 3, "Animation 4", "output", 0, "hit3", "aiming", 0, "state", "aiming", 1, "aim", "hit1", 0, "aiming", "hit1", 1, "Animation 5", "hit2", 0, "hit1", "hit2", 1, "Animation 6", "hit3", 0, "hit2", "hit3", 1, "Animation 7" ]
+node_connections = [ "state", 0, "Animation", "state", 1, "Animation 3", "state", 2, "Animation 2", "state", 3, "Animation 4", "output", 0, "hit3", "hit3", 0, "hit2", "hit3", 1, "Animation 7", "hit2", 0, "hit1", "hit2", 1, "Animation 6", "hit1", 0, "aiming", "hit1", 1, "Animation 5", "aiming", 0, "state", "aiming", 1, "aim" ]
 
 [sub_resource type="SphereShape" id=53]
 radius = 1.11815
@@ -770,18 +770,32 @@ script = ExtResource( 1 )
 [node name="Skeleton" parent="RedRobotModel/Armature" index="0"]
 bones/14/bound_children = [ NodePath("RayFrom") ]
 
+[node name="EvilRoboBody" parent="RedRobotModel/Armature/Skeleton" index="0"]
+layers = 5
+
+[node name="EvilRoboBody001" parent="RedRobotModel/Armature/Skeleton" index="1"]
+layers = 5
+
+[node name="EvilRoboShield-L" parent="RedRobotModel/Armature/Skeleton" index="2"]
+layers = 5
+
+[node name="EvilRoboShield-R" parent="RedRobotModel/Armature/Skeleton" index="3"]
+layers = 5
+
 [node name="RayFrom" type="BoneAttachment" parent="RedRobotModel/Armature/Skeleton" index="4"]
-transform = Transform( -0.997439, -0.070985, -0.00878045, -0.0710157, 0.99747, 0.00323016, 0.00852894, 0.00384544, -0.999956, -0.149183, 1.9902, 0.191941 )
+transform = Transform( -0.997632, -0.0685251, -0.00582601, -0.0685744, 0.997608, 0.00872382, 0.00521427, 0.00910268, -0.999945, -0.143299, 2.01235, 0.186807 )
 bone_name = "CannonAnimRecoil"
 
 [node name="Circle" type="MeshInstance" parent="RedRobotModel/Armature/Skeleton/RayFrom"]
 transform = Transform( 1, 0, 1.73472e-18, 0, -8.88178e-16, -1, 0, 1, 0, 0, 0, -1 )
+layers = 5
 mesh = SubResource( 2 )
 skeleton = NodePath("")
 material/0 = ExtResource( 4 )
 
 [node name="RayMesh" type="MeshInstance" parent="RedRobotModel/Armature/Skeleton/RayFrom"]
 transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, -1.25 )
+layers = 5
 mesh = SubResource( 4 )
 skeleton = NodePath("")
 material/0 = ExtResource( 4 )
@@ -1105,6 +1119,12 @@ unit_size = 15.0
 stream = ExtResource( 19 )
 unit_db = 3.0
 unit_size = 15.0
+
+[node name="FillLight" type="OmniLight" parent="."]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 2.5, 1.4 )
+light_energy = 0.6
+light_cull_mask = 4293918724
+omni_range = 4.0
 
 [connection signal="body_entered" from="PlayerDetectionArea" to="." method="_on_area_body_entered"]
 [connection signal="body_exited" from="PlayerDetectionArea" to="." method="_on_area_body_exited"]