--- a/level/forklift/mid_cargo_box.tres
+++ b/level/forklift/mid_cargo_box.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "mid_cargo_box"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 2 )
 metallic = 1.0
 metallic_texture = ExtResource( 3 )