--- a/level/geometry/models/container2.tres
+++ b/level/geometry/models/container2.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "container2"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 3 )
 metallic = 0.5
 metallic_texture = ExtResource( 1 )