--- a/level/geometry/models/cycles_glow_orange.tres
+++ b/level/geometry/models/cycles_glow_orange.tres
@@ -2,5 +2,6 @@
 
 [resource]
 resource_name = "cycles_glow_orange"
+params_diffuse_mode = 1
 albedo_color = Color( 0.639282, 0.251191, 0.0284811, 1 )
 roughness = 0.0960784