--- a/level/geometry/models/door.tres
+++ b/level/geometry/models/door.tres
@@ -7,6 +7,7 @@
 
 [resource]
 resource_name = "door"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 4 )
 metallic = 0.5
 metallic_texture = ExtResource( 2 )