--- a/level/geometry/models/flat_emitter_blue.tres
+++ b/level/geometry/models/flat_emitter_blue.tres
@@ -2,6 +2,8 @@
 
 [resource]
 resource_name = "flat_emitter_blue"
+params_diffuse_mode = 1
+params_specular_mode = 4
 albedo_color = Color( 0, 0.612694, 0.64, 1 )
 roughness = 0.542695
 emission_enabled = true