--- a/level/geometry/models/flat_emitter_red.tres
+++ b/level/geometry/models/flat_emitter_red.tres
@@ -2,5 +2,7 @@
 
 [resource]
 resource_name = "flat_emitter_red"
+params_diffuse_mode = 1
+params_specular_mode = 4
 albedo_color = Color( 0.64, 0, 0.00179112, 1 )
 roughness = 0.0960784