--- a/level/geometry/models/flat_emitter_white.tres
+++ b/level/geometry/models/flat_emitter_white.tres
@@ -2,6 +2,8 @@
 
 [resource]
 resource_name = "flat_emitter_white"
+params_diffuse_mode = 1
+params_specular_mode = 4
 albedo_color = Color( 0.64, 0.64, 0.64, 1 )
 roughness = 0.542695
 emission_enabled = true