--- a/level/geometry/models/mid_cargo_box.tres
+++ b/level/geometry/models/mid_cargo_box.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "mid_cargo_box"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 1 )
 metallic = 0.5
 metallic_texture = ExtResource( 2 )