--- a/level/geometry/models/reactor_cylinder.tres
+++ b/level/geometry/models/reactor_cylinder.tres
@@ -7,6 +7,7 @@
 
 [resource]
 resource_name = "reactor_cylinder"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 3 )
 metallic = 0.5
 metallic_texture = ExtResource( 4 )