--- a/level/geometry/models/reactor_emitter.tres
+++ b/level/geometry/models/reactor_emitter.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "reactor_emitter"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 1 )
 metallic = 0.5
 metallic_texture = ExtResource( 2 )