--- a/level/geometry/models/scifi_car_glass.tres
+++ b/level/geometry/models/scifi_car_glass.tres
@@ -3,6 +3,7 @@
 [resource]
 resource_name = "scifi_car_glass"
 flags_transparent = true
+params_diffuse_mode = 1
 albedo_color = Color( 0.191406, 0.0710055, 0.0710055, 0.720039 )
 metallic = 1.0
 roughness = 0.0