--- a/level/geometry/models/set_asphalt.tres
+++ b/level/geometry/models/set_asphalt.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "set_asphalt"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 1 )
 metallic = 0.5
 metallic_texture = ExtResource( 3 )