--- a/level/geometry/models/tile_anti_slip_metal.tres
+++ b/level/geometry/models/tile_anti_slip_metal.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "tile_anti_slip_metal"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 1 )
 metallic = 1.0
 metallic_texture = ExtResource( 3 )