--- a/level/geometry/models/tile_carbon_fiber.tres
+++ b/level/geometry/models/tile_carbon_fiber.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "tile_carbon_fiber"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 2 )
 metallic = 0.5
 metallic_texture = ExtResource( 1 )