--- a/level/geometry/models/tile_default_metal.tres
+++ b/level/geometry/models/tile_default_metal.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "tile_default_metal"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 2 )
 metallic = 0.5
 metallic_texture = ExtResource( 3 )