--- a/level/geometry/models/tile_glass.tres
+++ b/level/geometry/models/tile_glass.tres
@@ -8,6 +8,7 @@
 [resource]
 resource_name = "tile_glass"
 flags_transparent = true
+params_diffuse_mode = 1
 albedo_color = Color( 0.0120779, 0.0427238, 0.0474042, 0.43502 )
 albedo_texture = ExtResource( 1 )
 metallic = 0.5