--- a/level/geometry/models/tile_metal_pillars.tres
+++ b/level/geometry/models/tile_metal_pillars.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "Tile_metal_pillars"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 1 )
 metallic = 0.5
 metallic_texture = ExtResource( 3 )