--- a/level/geometry/models/tile_painted_gun_metal.tres
+++ b/level/geometry/models/tile_painted_gun_metal.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "tile_painted_gun_metal"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 2 )
 metallic = 0.5
 metallic_texture = ExtResource( 3 )