--- a/level/geometry/models/tile_silver_metal.tres
+++ b/level/geometry/models/tile_silver_metal.tres
@@ -2,6 +2,7 @@
 
 [resource]
 resource_name = "tile_silver_metal"
+params_diffuse_mode = 1
 albedo_color = Color( 0.21507, 0.216532, 0.239556, 1 )
 metallic = 0.5
 roughness = 0.0960784