--- a/level/geometry/models/tile_tech_panels_color.tres
+++ b/level/geometry/models/tile_tech_panels_color.tres
@@ -6,6 +6,7 @@
 
 [resource]
 resource_name = "tile_tech_panels_color"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 3 )
 metallic = 0.5
 metallic_texture = ExtResource( 2 )