--- a/level/geometry/models/train_inner.tres
+++ b/level/geometry/models/train_inner.tres
@@ -2,6 +2,8 @@
 
 [resource]
 resource_name = "train_inner"
+params_diffuse_mode = 1
+params_specular_mode = 4
 albedo_color = Color( 0.10375, 0.125224, 0.213956, 1 )
 metallic = 0.5
 roughness = 0.0960784