--- a/level/geometry/models/trim_emission_lights.tres
+++ b/level/geometry/models/trim_emission_lights.tres
@@ -7,6 +7,7 @@
 
 [resource]
 resource_name = "trim_emission_lights"
+params_diffuse_mode = 1
 albedo_texture = ExtResource( 2 )
 metallic = 0.5
 metallic_texture = ExtResource( 1 )