--- a/level/level.tscn
+++ b/level/level.tscn
@@ -1,8 +1,8 @@
-[gd_scene load_steps=15 format=2]
+[gd_scene load_steps=16 format=2]
 
 [ext_resource path="res://level/level.gd" type="Script" id=1]
 [ext_resource path="res://level/geometry/scenes/props.tscn" type="PackedScene" id=2]
-[ext_resource path="res://level/geometry/giprobe_data.res" type="GIProbeData" id=5]
+[ext_resource path="res://level/level.lmbake" type="BakedLightmapData" id=3]
 [ext_resource path="res://player/player.tscn" type="PackedScene" id=6]
 [ext_resource path="res://enemies/red_robot/red_robot.tscn" type="PackedScene" id=7]
 [ext_resource path="res://level/level_music.ogg" type="AudioStream" id=8]
@@ -34,6 +34,9 @@ tracks/0/keys = {
 "values": [ Vector3( 0, 0, 0 ), Vector3( 0, -360, 0 ) ]
 }
 
+[sub_resource type="OccluderShapeSphere" id=3]
+spheres = [ Plane( 0, -22.326, 0, 11.661 ), Plane( 52.489, -7.5, 75.657, 4.5 ), Plane( 36.628, -4.828, 92.5, 4 ), Plane( -17.415, -4.017, 88.384, 12 ), Plane( 69.656, -2.014, 71, 4 ), Plane( 85.091, -11.5, 6.42, 8 ), Plane( 27.881, -13.9, 8.221, 8 ), Plane( -21.687, -13.9, 19.99, 8 ), Plane( -5.906, -13.9, -28.228, 8 ) ]
+
 [node name="Level" type="Spatial"]
 script = ExtResource( 1 )
 
@@ -48,17 +51,7 @@ script = ExtResource( 1 )
 [node name="WorldEnvironment" type="WorldEnvironment" parent="."]
 environment = ExtResource( 16 )
 
-[node name="GIProbe" type="GIProbe" parent="."]
-transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, -6.09263, 1.28266, 2.88598 )
-subdiv = 2
-extents = Vector3( 113.946, 48.8048, 117.028 )
-energy = 2.2
-propagation = 1.0
-interior = true
-data = ExtResource( 5 )
-
 [node name="ReflectionProbes" type="Spatial" parent="."]
-visible = false
 
 [node name="ReflectionProbe1" type="ReflectionProbe" parent="ReflectionProbes"]
 transform = Transform( 0.494055, 0, -0.869431, 0, 1, 0, 0.869431, 0, 0.494055, 28.1433, -18.8082, 71.0693 )
@@ -85,6 +78,18 @@ interior_ambient_contrib = 0.3
 
 [node name="Player" parent="." instance=ExtResource( 6 )]
 transform = Transform( -0.575826, 0, -0.817573, 0, 1, 0, 0.817573, 0, -0.575826, 64.8183, -1.0765, 78.7639 )
+
+[node name="FillLight" type="OmniLight" parent="Player"]
+transform = Transform( 0.999999, 0, 0, 0, 1, 0, 0, 0, 0.999999, -0.167068, 1.7, -1.10553 )
+light_energy = 0.1
+light_bake_mode = 0
+light_cull_mask = 4293918722
+
+[node name="FillLight2" type="OmniLight" parent="Player"]
+transform = Transform( 0.999999, 0, 0, 0, 1, 0, 0, 0, 0.999999, 0.178429, 0.6, -0.614975 )
+light_energy = 0.1
+light_bake_mode = 0
+light_cull_mask = 4293918722
 
 [node name="RedRobot1" parent="." instance=ExtResource( 7 )]
 transform = Transform( 0.843905, 0, -0.536493, 0, 1, 0, 0.536493, 0, 0.843905, 71.5907, -6.05686, 46.2736 )
@@ -134,7 +139,6 @@ autoplay = "mawaru"
 anims/mawaru = SubResource( 2 )
 
 [node name="Debug" type="Label" parent="."]
-visible = false
 margin_left = 10.0
 margin_top = 10.0
 margin_right = 40.0
@@ -146,3 +150,15 @@ script = ExtResource( 9 )
 __meta__ = {
 "_edit_use_anchors_": false
 }
+
+[node name="Occluder" type="Occluder" parent="."]
+shape = SubResource( 3 )
+
+[node name="BakedLightmap" type="BakedLightmap" parent="."]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, -6, 1, 3 )
+extents = Vector3( 113, 48, 117 )
+quality = 3
+bounce_indirect_energy = 3.0
+use_hdr = false
+capture_enabled = false
+light_data = ExtResource( 3 )