--- a/menu/settings.gd
+++ b/menu/settings.gd
@@ -1,11 +1,5 @@
 extends Node
 
-enum GIQuality {
-	DISABLED = 0,
-	LOW = 1,
-	HIGH = 2,
-}
-
 enum AAQuality {
 	DISABLED = 0,
 	AA_2X = 1,
@@ -32,7 +26,6 @@ enum Resolution {
 	NATIVE = 3,
 }
 
-var gi_quality = GIQuality.LOW
 var aa_quality = AAQuality.AA_2X
 var ssao_quality = SSAOQuality.DISABLED
 var bloom_quality = BloomQuality.HIGH
@@ -60,9 +53,6 @@ func load_settings():
 	if typeof(d) != TYPE_DICTIONARY:
 		return
 
-	if "gi" in d:
-		gi_quality = int(d.gi)
-
 	if "aa" in d:
 		aa_quality = int(d.aa)
 
@@ -84,5 +74,5 @@ func save_settings():
 	var error = f.open("user://settings.json", File.WRITE)
 	assert(not error)
 
-	var d = { "gi":gi_quality, "aa":aa_quality, "ssao":ssao_quality, "bloom":bloom_quality, "resolution":resolution, "fullscreen":fullscreen }
+	var d = { "aa":aa_quality, "ssao":ssao_quality, "bloom":bloom_quality, "resolution":resolution, "fullscreen":fullscreen }
 	f.store_line(to_json(d))