--- a/player/bullet/effect/BlastMaterial.tres
+++ b/player/bullet/effect/BlastMaterial.tres
@@ -4,7 +4,7 @@
 
 [sub_resource type="Shader" id=1]
 code = "shader_type spatial;
-render_mode blend_mix,depth_draw_never,cull_back,diffuse_burley,specular_schlick_ggx;
+render_mode blend_mix,depth_draw_never,cull_back,diffuse_lambert,specular_disabled;
 uniform vec4 albedo : hint_color;
 uniform sampler2D texture_albedo : hint_albedo;
 uniform float specular;