--- a/player/textures/player_robot_emissive.png.import
+++ b/player/textures/player_robot_emissive.png.import
@@ -29,6 +29,7 @@ process/fix_alpha_border=true
 process/premult_alpha=false
 process/HDR_as_SRGB=false
 process/invert_color=false
+process/normal_map_invert_y=false
 stream=false
 size_limit=0
 detect_3d=false