… effects_shared/BlastMesh_Sphere.mesh
… enemies/red_robot/laser/BarrelSmoke_Plane.mesh
… enemies/red_robot/laser/impact_effect/LightRaysMesh_Plane.mesh
… enemies/red_robot/parts/sparks_effect/SparkParticle_Sphere.mesh
… level/geometry/environment.tres
… level/geometry/models/core.glb.import
… level/geometry/models/lights.glb.import
… level/geometry/models/lights.glb.unwrap_cache
… level/geometry/models/props.glb.import
… level/geometry/models/structure.glb.import
… level/level.lmbake
… player/model/muzzle_flash_particle_mesh_Cone.mesh
… player/player.tscn
… project.godot