**Don't merge before 3.4 is released, as this relies on occluder spheres and other 3.4-only features.**

- Replace GIProbe with BakedLightmap. This alone speeds up scene rendering by at least 2× (if not more). The scene is lit in a different way (brighter overall now), but this also helps make it more readable.
  - Lightmap object capture is disabled as it would have created a file that's too large to be committed on GitHub (which imposes a 20 MB per file limit).
  - HDR is disabled on the lightmap to reduce its file size below 20 MB. Clipping or banding isn't really visible since only indirect lighting is baked anyway.
  - Reflection probes are used instead to provide reflections and improve the appearance of dynamic objects.
  - Reduce SSAO intensity to compensate for the lighting being brighter on average.
- Add fill lights to dynamic objects with a cull mask set to only impact themselves.
- Use the cheaper Lambert diffuse mode for materials.
- Set specular mode to Disabled on materials that don't need specular lighting.
- Disable shadows for most lights (they made little visual difference). This divides draw calls by 3 on average.
- Add occluder spheres. Gains are minor due to GLES3 already having a depth prepass and the level only having a few occlusion culling opportunities, but they're present in certain situations.

The GI setting was removed as it's no longer relevant.

Disabling shadows for non-essential lights and adding occluder spheres brings down the number of draw calls from 10k-15k to 2k-6k. (Most of the gains are from disabling shadows.)

## Performance

**CPU:** Intel Core i7-6700K
**GPU:** GeForce GTX 1080 (NVIDIA 470.74)
**OS:** Fedora 34

Performance measured at start area (after starting a game and waiting a few seconds):

| Settings | Before | After |
|----------|--------|-------|
| High settings* | 60 FPS (16.7 mspf) | 118 FPS (8.5 mspf) |
| Low settings* | 135 FPS (7.4 mspf) | 200 FPS (5 mspf) |

*High settings:* 2560×1440, MSAA 4×, AO High, Glow High, GI High (if relevant)
*Low settings:* 2560×1440, MSAA Off, AO Off, Glow Off, GI Off (if relevant)

Note that in certain areas of the level, the performance gains are even greater. With AO disabled but Glow kept to high, it's uncommon to go below 144 FPS during gameplay now. This makes for a much better experience on a 144 Hz monitor :slightly_smiling_face: 

## Preview

*Some visual cutbacks were made, but they usually aren't noticeable unless you go looking for them.*

![2021-10-13_19 04 08](https://user-images.githubusercontent.com/180032/137183021-0906ae1f-f33e-4291-9c43-82e00a8ee752.png)

![2021-10-13_19 04 18](https://user-images.githubusercontent.com/180032/137183026-ad8369cd-fb98-4d5b-82a5-af8862a3f8f4.png)

![2021-10-13_19 03 36](https://user-images.githubusercontent.com/180032/137183017-7017e034-0cf3-4dc0-9db2-3c59e7e94a80.png)

### Before

![2021-10-17_17 26 39](https://user-images.githubusercontent.com/180032/137634154-9b97cea6-9f8f-423e-bd48-18a9436e54a6.png)

### After

![2021-10-17_17 27 28](https://user-images.githubusercontent.com/180032/137634164-5e7de8bf-ebc9-4be7-9e99-f279dbe3dc8b.png)