@aaronfranke 2021-10-16
Both the before and after screenshots in your latest posts are *massively* brighter than the current version (and, full disclaimer, I think this looks significantly worse). I think that the current aesthetic of the demo is good and shouldn't be massively changed.

Significantly changing the appearance of the demo also really shouldn't be tied to improving performance. If we have to make the demo look significantly worse to improve performance, is that actually an improvement?