@aaronfranke 2021-10-16
Here is what I see after building the latest 3.x, clearing `.import/`, and opening this PR: https://imgsli.com/NzcyNjk

I can reproduce the bright ground, but it still looks significantly worse.