@Calinou 2021-10-17
> What's the purpose behind making the _visual_ demo more performant? Isn't the whole point to show the graphical capabilities of the engine? Why actively make the demo look worse to improve its performance? That seems kind of backwards to me. Especially since it's just a demo, not a template or asset pack or anything like that.
> 
> And if the argument is so that users on lower-end hardware can see the demo... it's a _visual_ demo. They can watch videos and get the same point across.

The demo is currently not quite smooth on high-end hardware from a few years ago. A GTX 1080 is still significantly faster than the [most popular GPU on Steam as of writing, which is a GTX 1060](https://store.steampowered.com/hwsurvey).

This gives a bad image of Godot to people who try it on mid-range hardware and dismiss it as a low-performance option compared to other engines.