@Calinou 2021-10-17
> Why not upload it as a separate demo? something like "TPS demo - performance optimized"?

To avoid confusion, we can't have too many similar-looking demos lying around.