@Nukiloco 2021-10-17
Looking at the after images, the entire quality seem to be much worse than it was before. I'm also going to recommend not to merge or change this at all and keep it the same.
In my opinion this demo is supposed to show how good Godot is at visual stuff, lowering down the quality of it is going to make it look worse and might get some new comers not interested at all.

>
> This gives a bad image of Godot to people who try it on mid-range hardware, and then dismiss it as a low-performance option > compared to other engines. This is something I've seen happen time and time again on various community platforms (inside and outside active Godot users). Overall, I've seen more people complain about Godot's poor 3D performance rather than poor visuals.

The engine in this case is what needs to be optimized here. Lowering down the quality of the game will make the situation even worse than it is before and doesn't fix the overall problem that your explaining here.

In my opinion and I think that we as a community, should take step back and look at the whole graphical part of the engine to see what can be optimized and fixed before making any changes to the demo itself. When I think of optimizations, I don't think of major graphical stuff that shows the engine's potential being cut just to fix a problem.