@Nukiloco 2021-10-17
And even if those engine optimizations aren't enough, I think a better solution is to bring in graphical options into the demo to allow the user to toggle between high and low graphics for computers with low end or high end specs.