@Jamsers 2021-11-09
Why does switching from GI Probe to Baked Lightmap reduce quality so much? Shouldn't Baked Lightmap actually produce higher quality results, albeit at a bigger cost to storage? I've tried it myself with my own copy of the TP demo and it does produce flat, bad lighting