@Calinou 2021-11-09
> Why does switching from GI Probe to Baked Lightmap reduce quality so much? Shouldn't Baked Lightmap actually produce higher quality results, albeit at a bigger cost to storage? I've tried it myself with my own copy of the TP demo and it does produce flat, bad lighting

Baked lightmaps in 3.x only store diffuse lighting, not specular lighting. They also don't store directional information in 3.x, although GIProbes don't do that either for indirect lighting.

Baked lightmaps are also much less prone to leaks, but GIProbe leaking is barely noticeable in the TPS demo since it's an indoor scene.